[FIX] 워크스페이스 삭제 에러 해결

+++ module-user/src/main/java/com/xiilab/moduleuser/service/GroupServiceImpl.java
@@ -86,13 +86,17 @@ public class GroupServiceImpl implements GroupService {
 	public void deleteWorkspaceGroupByName(String groupName) {
 		GroupInfoDTO groupInfo = groupRepository.getGroupInfoByCategoryAndName(WORKSPACE,
 			groupName);
-		groupRepository.deleteGroupById(groupInfo.getUid());
+		if (groupInfo != null) {
+			groupRepository.deleteGroupById(groupInfo.getUid());
+		}
 	}
 
 	@Override
 	public void deleteAccountGroupByName(String groupName) {
 		GroupInfoDTO groupInfo = groupRepository.getGroupInfoByCategoryAndName(ACCOUNT, groupName);
-		groupRepository.deleteGroupById(groupInfo.getUid());
+		if (groupInfo != null) {
+			groupRepository.deleteGroupById(groupInfo.getUid());
+		}
 	}
 
 	private void createWorkspaceChildGroup(String parentId, GroupReqDTO groupReqDTO, UserInfoDTO userInfoDTO) {
